# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kube API service is service for connection with Kubernetes.",
    "title": "Kube API"
  },
  "paths": {
...
    "/ingresses": {
      "get": {
        "tags": [
          "Ingress"
        ],
        "summary": "Get ingresses from all user namespaces.",
        "operationId": "GetSelectedIngresses",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          }
        ],
        "responses": {
          "200": {
            "description": "ingresses list from all users namespaces",
            "schema": {
              "$ref": "#/definitions/SelectedIngressesList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      }
    },
...
    "/namespaces/{namespace}": {
      "get": {
        "tags": [
          "Namespace"
        ],
        "summary": "Get namespace.",
        "operationId": "GetNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "namespace",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "Namespace"
        ],
        "summary": "Update namespace.",
        "operationId": "UpdateNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "namespace updated",
            "schema": {
              "$ref": "#/definitions/NamespaceWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      },
      "delete": {
        "tags": [
          "Namespace"
        ],
        "summary": "Delete namespace.",
        "operationId": "DeleteNamespace",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "namespace deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      }
    },
...
    "/namespaces/{namespace}/ingresses": {
      "get": {
        "tags": [
          "Ingress"
        ],
        "summary": "Get ingresses list.",
        "operationId": "GetIngressList",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ingresses list",
            "schema": {
              "$ref": "#/definitions/IngressesList"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "post": {
        "tags": [
          "Ingress"
        ],
        "summary": "Create ingress.",
        "operationId": "CreateIngress",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/IngressWithOwner"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ingress created",
            "schema": {
              "$ref": "#/definitions/IngressWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      }
    },
    "/namespaces/{namespace}/ingresses/{ingress}": {
      "get": {
        "tags": [
          "Ingress"
        ],
        "summary": "Get ingresses list.",
        "operationId": "GetIngress",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ingress",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ingresses",
            "schema": {
              "$ref": "#/definitions/IngressWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      },
      "put": {
        "tags": [
          "Ingress"
        ],
        "summary": "Update ingress.",
        "operationId": "UpdateIngress",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ingress",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/IngressWithOwner"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "ingress updated",
            "schema": {
              "$ref": "#/definitions/IngressWithOwner"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      },
      "delete": {
        "tags": [
          "Ingress"
        ],
        "summary": "Delete ingress.",
        "operationId": "DeleteIngress",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserNamespaceHeader"
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ingress",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "ingress deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "private"
      }
    },
...
  },
  "definitions": {
...
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/cherry"
    },
    "Ingress": {
      "description": "Ingress -- model for ingress",
      "type": "object",
      "required": [
        "name",
        "rules"
      ],
      "properties": {
        "created_at": {
          "description": "creation date in RFC3339 format",
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "rules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Rule"
          },
          "x-go-name": "Rules"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "IngressesList": {
      "description": "IngressesList -- model for ingresses list",
      "type": "object",
      "properties": {
        "ingresses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ingress"
          },
          "x-go-name": "Ingress"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Path": {
      "description": "Path -- ingress path",
      "type": "object",
      "required": [
        "path",
        "service_name",
        "service_port"
      ],
      "properties": {
        "path": {
          "type": "string",
          "x-go-name": "Path"
        },
        "service_name": {
          "type": "string",
          "x-go-name": "ServiceName"
        },
        "service_port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ServicePort"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Rule": {
      "description": "Rule -- ingress rule",
      "type": "object",
      "required": [
        "host",
        "path"
      ],
      "properties": {
        "host": {
          "type": "string",
          "x-go-name": "Host"
        },
        "path": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Path"
          },
          "x-go-name": "Path"
        },
        "tls_secret": {
          "type": "string",
          "x-go-name": "TLSSecret"
        }
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "SelectedIngressesList": {
      "description": "SelectedIngressesList -- model for ingresses list from all namespaces",
      "type": "object",
      "additionalProperties": {
        "$ref": "#/definitions/IngressesList"
      },
      "x-go-package": "git.containerum.net/ch/kube-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
  },
  "parameters": {
...
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserNamespaceHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserNamespacesHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}